Support Redis Cluster configured with a single node
This patch adds support for a Redis Cluster that's configured with only
a single initial host. This is required to make this gem play nicely
with AWS ElastiCache Redis which suggests configuring your clients with
just a single "configuration endpoint" that's used to return the rest of
the active node configuration.

Currently when configured with just one host, the gem short circuits in
`#reload_pool_nodes` after initializing the pool with that one node
handling the cluster's entire slot space. As it stands when it's run
against an ElastiCache Redis, the loop in `#execute` fails again and
again as it gets a `MOVE` response from the cluster, but never corrects
itself because `#reload_pool_nodes` will never make a `CLUSTER SLOTS`
request. Instead, the loop's TTL eventually expires and it falls through
with a mysterious `false` return.

This patch aims to support three distinct modes of operation:

1. A standalone (non-Cluster Redis).
2. A Redis Cluster configured with only one node (i.e., ElastiCache).
3. A Redis Cluster configured with multiple nodes (presumably the most
   commonly seen cluster configuration).

To that end you can now, you can now initialize `Redis::Client` two
different ways:

1. With a `Hash` containing a single host/port combination. In this case
   the library will always assume that the client wants a standalone
   configuration.
2. An `Array` of host/port combinations. This is the same as before,
   except an array of length one is now supported and does the right
   thing.

One last addition is a new `force_cluster` configuration option. When
set to `false`, it causes the library to revert to standalone operation
even if it was configured for Redis Cluster. This is convenient for test
suites where you might want to talk to a standalone Redis in testing and
development, but a cluster in production or other cloud environments.