Introduce `#reconnect` method
Introduces a `#reconnect` method that allows users of the library to get
finer grain control over how it behaves. When received, `#reconnect`
closes every connection that it knows about and reloads it pool. This is
useful, for example, to force a reconnect when receiving certain types
of errors.

`#reconnect` has an optional argument `use_initial_hosts` that tells the
client to restart from the object's original configuration instead of
reusing whatever it's holding from the last time it requested `CLUSTER
SLOTS`. This should be useful for some users who want to guarantee a
clean slate.