Introduce basic logging facilities
When things go wrong it's nice to have a little more insight into what
exactly any given library is doing. This patch introduces some basic
(and optional) logging facilities to help with this.

The most minimal approach is to enable the logger at the error level. In
this case the library will just log errors that it sees and handles.