Add more detailed error messages and improve naming
While developing a project with `redis_cluster` initially, I found
debugging some parts to be quite difficult because the errors thrown by
the library don't have messages and as a result, are quite vague. This
patch adds some explanatory messages to each error class which will
hopefully be helpful to users who are trying to debug.

A little more contentious is that I renamed a couple of the classes so
that they're more grammatically correct. I've added backwards-compatible
type aliases for them so that this isn't a breaking change, but those
could plausibly be dropped on the next major version increment.