Refactor how the library retries failed calls
The original impetus was some fragility around how
`#create_multi_node_pool` attempts to repopulate its pool: it would
cycle through each of its seed hosts and execute a command against it.
If it ran into an intermittent connection problem, it would swallow
exceptions along the way.

This is usually fine if you have a reasonably-sized seed pool and
intermittent connections really are intermittent (and not common), but
it breaks down in cases like ElastiCache Redis Cluster where you may
only have a seed host. The library would fail on that host, drop out of
the loop, and leave itself with a pool initialized to zero nodes.
Later, as a call was executed on it, it would bubble a runtime exception
up the chain to the effect of `#execute` called on `nil`.

I went in to rework `#create_multi_node_pool` so that instead of cycling
through each host sequentially, we'd instead `#sample` one of the seeds
randomly and then retry intermittent failures a little like how the main
`#execute` loop did.

That's done, but as I was extracting a common helper for the two methods
to share (`#retry_intermittent_loop`), I realized that we could improve
the way that `#execute` handles error as well. Its loop was treating all
exceptions that came out of an execution whether they be an intermittent
problem, `ASK`/`MOVED`, or other command error equally. This is
undesirable for a few reasons:

1. `ASK`/`MOVED` is really more like a flow control response than an
   exception. Ideally, you'd receive it once, retry it once, and you're
   away.

2. Errors that cannot ever be plausibly be corrected (like
   non-`ASK`/`MOVED` and non-auth `CommandError`s were being retried.

3. Library bugs or a misbehaving server could lead to `MOVED` being
   retried over and over again with the results not changing (I've
   experienced this previously).

By refactoring so that there's two distinct types of potential retries
we can get better control over the program's flow and also more
correctness. Commands now retry intermittent failures both during normal
`#execute` but also when reloading the node pool. An `ASK` or `MOVED`
will instigate a single retry, but no more. Other types of
`CommandError` are surfaced to the user immediately so that no work is
wasted on the side of either the client or server.

A side effect is that the `retry_count` configuration option is now more
sane: the number in here *just* applies to intermittent failures now, so
you don't have to try and account for possible `ASK`/`MOVED` operations.
Big shops like us can just set this number to `0` and have all
unexpected errors raised right out of the library so that we can handle
them ourselves.

I've tried to leave as many explanatory comments throughout the changes
as possible, but let me know if I can help with other information.