Use hash instead of optional parameters for Ruby 1.9 compatibility
Switches to an `options` hash instead of optional parameters so that we
can maintain compatibility with Ruby 1.9.

This approach is nice because when 1.9 support is dropped it's _mostly_
forward compatible if we then want to switch to optional parameters. It
does have the fairly sizable disadvantage though in that unlike optional
parameters, option names are not checked, so it's easy to misspell a
name like `use_initial_hosts` and not notice.

Addresses some feedback here:
https://github.com/zhchsf/redis_cluster/pull/15#issuecomment-367371837